pyproject.toml
[tool.poetry]
name = "oci-coordinator"
version = "0.1.0"
package-mode = false
description = "OCI AI Agent Coordinator - LangGraph orchestration for Oracle Cloud Infrastructure operations"
authors = ["OCI AI Team"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/yourusername/oci-coordinator"
documentation = "https://github.com/yourusername/oci-coordinator#readme"
keywords = ["oci", "oracle", "langgraph", "mcp", "ai-agents", "orchestration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"

# Core Frameworks
fastapi = ">=0.109.0"
uvicorn = {extras = ["standard"], version = ">=0.27.0"}
pydantic = ">=2.5.0"
pydantic-settings = ">=2.1.0"

# LangChain & LangGraph
langchain = ">=0.3.0"
langchain-core = ">=0.3.0"
langchain-community = ">=0.3.0"
langgraph = ">=0.2.0"
langgraph-checkpoint = ">=1.0.0"

# LLM Providers
langchain-anthropic = ">=0.2.0"
langchain-openai = ">=0.2.0"
# langchain-google-genai = ">=2.0.0"  # Optional: OCI GenAI

# OCI SDK
oci = ">=2.120.0"

# MCP Client
mcp = ">=1.0.0"

# Observability
opentelemetry-api = ">=1.22.0"
opentelemetry-sdk = ">=1.22.0"
opentelemetry-exporter-otlp = ">=1.22.0"
opentelemetry-instrumentation-fastapi = ">=0.44b0"
opentelemetry-instrumentation-httpx = ">=0.44b0"
structlog = ">=24.1.0"
prometheus-client = ">=0.19.0"

# HTTP Client
httpx = ">=0.26.0"
aiohttp = ">=3.9.0"

# Utilities
tenacity = ">=8.2.0"
python-dotenv = ">=1.0.0"
orjson = ">=3.9.0"
pyyaml = ">=6.0.0"

# Redis (for state persistence)
redis = ">=5.0.0"

# Channel Integrations
slack-sdk = ">=3.26.0"
slack-bolt = ">=1.18.0"
# botbuilder-core = ">=4.14.0"  # Optional: Microsoft Teams
fastmcp = "^2.14.1"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-asyncio = ">=0.23.0"
pytest-cov = ">=4.1.0"
pytest-mock = ">=3.12.0"
pytest-xdist = ">=3.5.0"
ruff = ">=0.2.0"
black = ">=24.1.0"
mypy = ">=1.8.0"
pre-commit = ">=3.6.0"
ipython = ">=8.20.0"

[tool.poetry.group.docs.dependencies]
mkdocs = ">=1.5.0"
mkdocs-material = ">=9.5.0"
mkdocstrings = {extras = ["python"], version = ">=0.24.0"}

[tool.poetry.scripts]
oci-coordinator = "src.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# ========================================
# Ruff Configuration
# ========================================
[tool.ruff]
target-version = "py311"
line-length = 88
fix = true
exclude = ["analysis"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "PTH",  # flake8-use-pathlib
    "ERA",  # eradicate (commented code)
    "PL",   # pylint
    "RUF",  # ruff-specific
]
ignore = [
    "E501",    # line too long (handled by black)
    "B008",    # do not perform function calls in argument defaults
    "PLR0913", # too many arguments
    "PLR2004", # magic value comparison
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "ARG", "PLR2004"]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

# ========================================
# Black Configuration
# ========================================
[tool.black]
target-version = ["py311"]
line-length = 88
extend-exclude = 'analysis/'

# ========================================
# MyPy Configuration
# ========================================
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "langgraph.*",
    "oci.*",
    "mcp.*",
    "slack_sdk.*",
    "slack_bolt.*",
]
ignore_missing_imports = true

# ========================================
# Pytest Configuration
# ========================================
[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# ========================================
# Coverage Configuration
# ========================================
[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
]
fail_under = 80
show_missing = true
